For the most part, ensure that only a single websocket connection is opened to the server.
Make sure that we can stream a transcript from the server to the
web interface.

Right now, we have some start and end words that are used to
identify what part of an endless stream of text should be
considered an instruction for the cobot.

The assumption is that audio will be transcribed to a pipe, from
which the codebotler application will read.

Tested by echoing text into the pipe.

Issues that will need to be fixed:
 - Tracking websocket connections and shutting down the server
   gracefully.
 - It seems like some of the added async routines are hanging or
   not responding to ctrl+c