Multiple changes:
 - Make shutdown actually close the blocking read on the audio_pipe
when Ctrl+C is pressed. However, the usage of global variables is
still present. Shutdown seems to occur cleanly now.

- Track clients that connect and disconnect from the websocket.
This permits us to stream transcripts to multiple connected clients

- Fixed so that we have only a single reader from the audio_pipe,
and whenever we read, it will be broadcasted to whatever clients
are connected.

- The last generated code is tracked. We now start executing code
when the audio request asks us to, rather than sending data to the
client to automate the execute request.